PUB-3 Refactor endpoints into a module
